requirements.txt
click==6.6
Flask==0.11.1
Flask-SocketIO==2.4
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
python-engineio==0.9.1
python-redmine==1.5.1
python-socketio==1.3
six==1.10.0
Werkzeug==0.11.10
